Successfully post a message.
I don't know why this wasn't an issue before, but for some
reason the pointer to the file handle used by the test client
was empty, so the form was failing even though the file
was valid. So now all calls that use a test audio file
should be wrapped in a with block so that they are closed
immediately after they are opened.