Improvements to AJAX posting audio to response form
In JavaScript the FormData cannot be overwritten easily,
so the recorded blob is actually sent given a different
name "content". Luckily, the CBV is smart enough to associate
the blob in request.FILES with the 'audio' field of the
ModelForm, but it means that validation of the audio
field has to come as a non-field-error.